+++ __tests__/spot/simple_earn/getCollateralRecord.test.js
@@ -1,26 +1,23 @@
 /* global describe, it, expect */
-const MissingParameterError = require('../../../src/error/missingParameterError')
 const { nockMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
-
 const { mockResponse } = require('../../testUtils/mockData')
 
-const productId = '1'
-
 describe('#getCollateralRecord', () => {
-  describe('throw MissingParameterError', () => {
-    it('missing productId', () => {
-      expect(() => {
-        SpotClient.getCollateralRecord('')
-      }).toThrow(MissingParameterError)
+  it('should return collateral records', () => {
+    nockMock('/sapi/v1/simple-earn/flexible/history/collateralRecord')(mockResponse)
+
+    return SpotClient.getCollateralRecord().then(response => {
+      expect(response).toBeDefined()
+      expect(response.data).toEqual(mockResponse)
     })
   })
-  it('should suscribe locked product', () => {
+  it('should return collateral records with params', () => {
     const parameters = {
-      productId
+      productId: '1'
     }
     nockMock(`/sapi/v1/simple-earn/flexible/history/collateralRecord?${buildQueryString(parameters)}`)(mockResponse)
 
-    return SpotClient.getCollateralRecord(productId).then(response => {
+    return SpotClient.getCollateralRecord(parameters).then(response => {
       expect(response).toBeDefined()
       expect(response.data).toEqual(mockResponse)
     })