+++ __tests__/spot/simple_earn/getFlexibleSubscriptionPreview.test.js
@@ -1,34 +1,33 @@
 /* global describe, it, expect */
 const MissingParameterError = require('../../../src/error/missingParameterError')
-const { nockPostMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
-
+const { nockMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
 const { mockResponse } = require('../../testUtils/mockData')
 
 const productId = '1'
 const amount = 10
 
-describe('#subscribeFlexibleProduct', () => {
+describe('#getFlexibleSubscriptionPreview', () => {
   describe('throw MissingParameterError', () => {
     it('missing productId', () => {
       expect(() => {
-        SpotClient.subscribeFlexibleProduct('', amount)
+        SpotClient.getFlexibleSubscriptionPreview('', amount)
       }).toThrow(MissingParameterError)
     })
 
     it('missing amount', () => {
       expect(() => {
-        SpotClient.subscribeFlexibleProduct(productId, '')
+        SpotClient.getFlexibleSubscriptionPreview(productId, '')
       }).toThrow(MissingParameterError)
     })
   })
-  it('should suscribe flexible product', () => {
+  it('should return flexible subscription previews', () => {
     const parameters = {
       productId,
       amount
     }
-    nockPostMock(`/sapi/v1/simple-earn/flexible/subscribe?${buildQueryString({ ...parameters })}`)(mockResponse)
+    nockMock(`/sapi/v1/simple-earn/flexible/subscriptionPreview?${buildQueryString({ ...parameters })}`)(mockResponse)
 
-    return SpotClient.subscribeFlexibleProduct(productId, amount).then(response => {
+    return SpotClient.getFlexibleSubscriptionPreview(productId, amount).then(response => {
       expect(response).toBeDefined()
       expect(response.data).toEqual(mockResponse)
     })