+++ __tests__/spot/simple_earn/getLockedSubscriptionPreview.test.js
@@ -1,34 +1,33 @@
 /* global describe, it, expect */
 const MissingParameterError = require('../../../src/error/missingParameterError')
 const { nockMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
-
 const { mockResponse } = require('../../testUtils/mockData')
 
-const productId = '1'
+const projectId = '1'
 const amount = 10
 
-describe('#getFlexibleSubscriptionPreview', () => {
+describe('#getLockedSubscriptionPreview', () => {
   describe('throw MissingParameterError', () => {
-    it('missing productId', () => {
+    it('missing projectId', () => {
       expect(() => {
-        SpotClient.getFlexibleSubscriptionPreview('', amount)
+        SpotClient.getLockedSubscriptionPreview('', amount)
       }).toThrow(MissingParameterError)
     })
 
     it('missing amount', () => {
       expect(() => {
-        SpotClient.getFlexibleSubscriptionPreview(productId, '')
+        SpotClient.getLockedSubscriptionPreview(projectId, '')
       }).toThrow(MissingParameterError)
     })
   })
-  it('should suscribe locked product', () => {
+  it('should return locked subscription preview', () => {
     const parameters = {
-      productId,
+      projectId,
       amount
     }
-    nockMock(`/sapi/v1/simple-earn/flexible/subscriptionPreview?${buildQueryString({ ...parameters })}`)(mockResponse)
+    nockMock(`/sapi/v1/simple-earn/locked/subscriptionPreview?${buildQueryString({ ...parameters })}`)(mockResponse)
 
-    return SpotClient.getFlexibleSubscriptionPreview(productId, amount).then(response => {
+    return SpotClient.getLockedSubscriptionPreview(projectId, amount).then(response => {
       expect(response).toBeDefined()
       expect(response.data).toEqual(mockResponse)
     })