+++ __tests__/spot/simple_earn/getRateHistory.test.js
@@ -1,26 +1,25 @@
 /* global describe, it, expect */
 const MissingParameterError = require('../../../src/error/missingParameterError')
-const { nockPostMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
-
+const { nockMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
 const { mockResponse } = require('../../testUtils/mockData')
 
 const productId = '1'
 
-describe('#redeemFlexibleProduct', () => {
+describe('#getRateHistory', () => {
   describe('throw MissingParameterError', () => {
     it('missing productId', () => {
       expect(() => {
-        SpotClient.redeemFlexibleProduct('')
+        SpotClient.getRateHistory('')
       }).toThrow(MissingParameterError)
     })
   })
-  it('should suscribe flexible product', () => {
+  it('should return rate history', () => {
     const parameters = {
       productId
     }
-    nockPostMock(`/sapi/v1/simple-earn/flexible/redeem?${buildQueryString({ ...parameters })}`)(mockResponse)
+    nockMock(`/sapi/v1/simple-earn/flexible/history/rateHistory?${buildQueryString(parameters)}`)(mockResponse)
 
-    return SpotClient.redeemFlexibleProduct(productId).then(response => {
+    return SpotClient.getRateHistory(productId).then(response => {
       expect(response).toBeDefined()
       expect(response.data).toEqual(mockResponse)
     })