+++ __tests__/spot/simple_earn/setLockedAutoSubscribe.test.js
@@ -1,15 +1,14 @@
 /* global describe, it, expect */
 const MissingParameterError = require('../../../src/error/missingParameterError')
 const { nockPostMock, buildQueryString, SpotClient } = require('../../testUtils/testSetup')
-
 const { mockResponse } = require('../../testUtils/mockData')
 
-const productId = '1'
+const positionId = '1'
 const autoSubscribe = true
 
 describe('#setLockedAutoSubscribe', () => {
   describe('throw MissingParameterError', () => {
-    it('missing productId', () => {
+    it('missing positionId', () => {
       expect(() => {
         SpotClient.setLockedAutoSubscribe('', autoSubscribe)
       }).toThrow(MissingParameterError)
@@ -17,18 +16,18 @@ describe('#setLockedAutoSubscribe', () => {
 
     it('missing autoSubscribe', () => {
       expect(() => {
-        SpotClient.setLockedAutoSubscribe(productId, '')
+        SpotClient.setLockedAutoSubscribe(positionId, '')
       }).toThrow(MissingParameterError)
     })
   })
-  it('should suscribe flexible product', () => {
+  it('should suscribe locked product', () => {
     const parameters = {
-      productId,
+      positionId,
       autoSubscribe
     }
     nockPostMock(`/sapi/v1/simple-earn/locked/setAutoSubscribe?${buildQueryString({ ...parameters })}`)(mockResponse)
 
-    return SpotClient.setLockedAutoSubscribe(productId, autoSubscribe).then(response => {
+    return SpotClient.setLockedAutoSubscribe(positionId, autoSubscribe).then(response => {
       expect(response).toBeDefined()
       expect(response.data).toEqual(mockResponse)
     })