+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 ## 3.3.0 - 2024-04-04
 
+### Added
+- Add Simple Earn endpoints:
+  - `GET /sapi/v1/simple-earn/flexible/list` to query available Simple Earn flexible product list
+  - `GET /sapi/v1/simple-earn/locked/list` to query available Simple Earn locked product list
+  - `POST /sapi/v1/simple-earn/flexible/subscribe` to subscribe to a flexible product
+  - `POST /sapi/v1/simple-earn/locked/subscribe` to subscribe to a locked product
+  - `POST /sapi/v1/simple-earn/flexible/redeem` to redeem a flexible product
+  - `POST /sapi/v1/simple-earn/locked/redeem` to redeem a locked product
+  - `GET /sapi/v1/simple-earn/flexible/position` to get a flexible product position
+  - `GET /sapi/v1/simple-earn/locked/position` to get a locked product position
+  - `GET /sapi/v1/simple-earn/account ` to get a simple account balances
+  - `GET /sapi/v1/simple-earn/flexible/history/subscriptionRecord` to get flexible subscription records
+  - `GET /sapi/v1/simple-earn/locked/history/subscriptionRecord ` to get locked subscription records
+  - `GET /sapi/v1/simple-earn/flexible/history/redemptionRecord ` to retrieve flexible redemption records
+  - `GET /sapi/v1/simple-earn/locked/history/redemptionRecord ` to retrieve locked redemption records
+  - `GET /sapi/v1/simple-earn/flexible/history/rewardsRecord ` to get flexible rewards history
+  - `GET /sapi/v1/simple-earn/locked/history/rewardsRecord ` to get locked rewards history
+  - `POST /sapi/v1/simple-earn/flexible/setAutoSubscribe` to set an auto-subscription to a flexible product
+  - `POST /sapi/v1/simple-earn/locked/setAutoSubscribe` to set an auto-subscription to a locked product
+  - `GET /sapi/v1/simple-earn/flexible/personalLeftQuota` to get flexible personal left quota
+  - `GET /sapi/v1/simple-earn/locked/personalLeftQuota` to get locked personal left quota
+  - `GET /sapi/v1/simple-earn/flexible/subscriptionPreview` to get flexible subscription preview
+  - `GET /sapi/v1/simple-earn/locked/subscriptionPreview` to get locked subscription previews
+  - `GET /sapi/v1/simple-earn/flexible/history/rateHistory` to get a rate history
+  - `GET /sapi/v1/simple-earn/flexible/history/collateralRecord` to get collateral records
+
 ### Changed
 - Update dependencies